requirements.txt
### Requirements
torchvision
torchaudio
lightning
librosa
soundfile
scipy
matplotlib
pandas
tensorflow
tensorflow-probability
torchmetrics
seaborn
ipykernel
thop
openl3
bs4
python-sofa
onnx
onnxruntime
torch_tb_profiler
ffmpegio
noisereduce
jupyterlab
tensorboard
torchlibrosa
timm
nnAudio
